Move simulation_util.py into util.py
