+++ simulation.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 import replay_log
-import simulation_util
+import util
 
 SOUCE_IP_INDEX =  2
 DESTINATION_IP_INDEX = 3
@@ -10,14 +10,14 @@ def run_simulation(request_log_file):
   # replaying request log
   start_time, end_time = replay_log.simulate_requests(request_log_file)
   # collect server logs
-  logs = get_server_logs(start_time, end_time)
+  logs = util.get_server_logs(start_time, end_time)
   # calculate the average latency
   latency_sum = 0
   request_count = 0
   for log in logs:
     client_ip = log[SOURCE_INDEX]
     server_ip = log[DESTINATION_INDEX]
-    distance = get_distance(client_ip, server_ip)
+    distance = util.get_distance(client_ip, server_ip)
     unit = 1000.0
     latency = distance / unit
     request_importance = 1
@@ -31,4 +31,3 @@ if __name__ == '__main__':
   average_latency = run_simulation('sample_log')
   print '************************* Average latency ****************************'
   print average_latency
-
